refactor guest placements
"Add a ""(Save Changes"")"" button above the ""Change to"" colunm. This will show a popup with:

""Changes to save:
>Table<

Email to send:
>Email<

(Save and don't email)   (Save and send emails)  ""

Table colunmns:
1) Family
2) Guests before change (># of guests: >Name 1< (>Gender<, >Age<), >Name 2... etc)
3) Guests after change (># of guests: >Name 1< (>Gender<, >Age<), >Name 2... etc)

Email:
Default subject text:
""Homestay placements for >Group Name< group"" (>From date< to >To date<) (Location: >location)

Default body text:
""Dear >parent 1 first name<,

Here is an update for the guest placements with you for the >Group Name< group (>From date< to >To date<) (Location: >location).

Here is the updated list of guests that are placed with you from this group:
|  Guest name  |   Aussie name |  Position  |  Age   | Gender   |
|                      |                        |               |           |              |

Here is a summary of your confirmed groups:
>Show ""Confiemed< groups table<

Here are other groups that you're applying for at this time:
>Show ""Applied< groups table<

Here are other available groups that you may want to apply for:
>Show ""Available< groups table<

Light regards,
The Choice Homestay Team"""