"When importing a list: 1) Don't require the date of birth 2) Enable ""Male"" and ""Female"", not only ""male"" and ""female""
Convert to show ""Male"" or ""Female"" on the guest list:
1) male -> Male
2) m  -> Male
3) M -> Male
4) female -> Female
5) f -> Female
6) F -> Female"
"Next to payments, show:
1) Individual: |                   |
2) 2-share: |                   |
3) 3-share: |                   |
4) 4-share: |                   |"
"Show:
""Refer family and friends that aren't with Choice yet and receive $30 the first time we place guests with them.
                            (Refer someone)

Existing referals:
[Table]
1) Email
2) First name
3) Family name
4) Mobile
5) Referral $
6) Hosted guests  (Automatic value: Yes/No)
7) Status (Not hosted / processing / paid)
[/Table] ""
----------------------------------------------
Clicking the ""(Refer someone)"" button will show a popup with:
""Please provide contact information for family or friends you want to refer, and we will contact them when we are next recruiting. If they're not already hosting with Choice, then you will receive $30 the first time we place guests with them.

Email:            [                     ]
First name:    [                     ]
Family name [                     ]
Mobile:          [                     ]
Landline:      [                      ]    ""
---------------------------------------------"
"1) Remove the ""Family Status"" colunm from this page
2) Remove the ""Family Sub status"" colunm from this page
3) Enable the ""Clear filter"" button to clear the ""Status"" filter options
4) Vertically centre the text in the fields"
"Add a colunm ""Updated"" and show the most recent date that:
5a. The blue card status was changed
5b. The ""Send"" button on the ""Apply or Link a blue card"" popup was clicked"
"When clicking on ""Apply of link a blue card"":
1) Reword the text to:
a. ""Do you want to apply for a new blue card or link an existing blue card?""
b. ""Apply for a new blue card""
c. ""Link an existing blue card""

2) Don't show the ""Uploaded at"" text unless there is a value to show here

3) Reword ""Add Upload the completed form"" to ""Upload another page"" and show this the same as the ""Choose files"" option, with a new button showing when a file is uploaded

4) Remove the ""File"" text above the browse option

5) Reword ""Choose Files"" to ""Choose file"""
"1) Reword ""Send"" to ""Save""
2) If the files haven't been provided, then they can still click ""Save""
3) If the files haven't been provided, then don't enable the ""[  ] My blue card application is ready to process"" option"
"Replace the download form files with these:
1) New Application: https://www.dropbox.com/s/o06pjasqf8gglaw/4%20Blue%20Card%20App.pdf
2) Linking Application: https://www.dropbox.com/s/c5cg5qrnt8gwi4l/4%20Blue%20Card%20Authorisation.pdf"
"When clicking on ""Apply of link a blue card"":
Once a family has uploaded the form and their 2 ID files, show:
""[  ] My blue card application is ready to process""

If they choose the ""[  ] My blue card application is ready to process"" option then update the blue card status to ""Ready to process"""
Reword "Sent" status to "Lodged"
"Don't show the ""Apply or linka  blue card"" button to families if their blue card status =
1) Approved
2) Exempt
3) Lodged"
"Make the height of the table more, so that the filter menu is visible:
https://www.dropbox.com/s/605j7zhp5cnv7z6/Screenshot%202017-04-12%2012.26.36.png"