"Add a ""Referrals"" section in the header, next to ""Emails"", and show all referrals here (the same data from the ""Family Profile -> Referrals""  page)
Add a colunm to this page called ""Profile Created"" and show the date their profile was created. Also add this ""Profile Created"" colunm to the ""Family Profile -> Referrals"" page "
When scrolling down the page, move the heading to it stays in view
"https://www.choicehomestay.com/group/list
The T11 group should be showing 4 guests not placed"
When a group had "0" guests not placed, show the "guests placed" value with green background
The "remove" red button is not showing on the family profile for pets and bedrooms for this person (Staff account: Maddison) https://www.choicehomestay.com/family/edit/8PSMciZRH45zsELpH
"Create a heading section called ""Guests"". Show all of the guest lists from all groups here, with:
1) Also show ""Group ID"" and ""Group Name"" colunm
2) Don't show the family address and mobile
3) Also show a colunm ""Date updated"" and show the date their placement with a family was updated""

The main reason for this table is to view the ""Email status"" for guest placements, to see if any are not ""confirmed"" for groups arriving soon, with the ""Date updated"" colunm there so we can be sure we provide a family with time after sending a confirmedion before we call them to confirm if they haven't clicked."
"Add a map view of the ""Applied"", ""Confirmed"" and ""Backup"" lists with:
1) list of family profiles on the left
2) google map on the right showng the study location and each family on the list

When moving the mouse over the family on the list or on the map, show the family listing and map position a different colour and show more information on the map position (family name, parents names, family status, family sub-status, address, phone)"